file-converter-app/requirements.txt
flask==2.3.3
python-docx==0.8.11
PyPDF2==3.0.1
pdf2image==1.16.3
Pillow==10.0.1
pandas==2.1.0
openpyxl==3.1.2
ffmpeg-python==0.2.0